backend/requirements.txt
fastapi
uvicorn
python-multipart

transformers
python-docx
PyPDF2
torch 